feat: send information from the request to sendVerificationRequest
Send more info to the sendVerificationRequest function that could be relevant to sending the email verification request.

In my case I wanted the user agent to have a message like "This login was requested using using Firefox on Windows" which can help a user verify that they are the one making the request.